refactor: simplify message type support functions
Remove inline implementations and enhance readability of message type support functions, while maintaining functionality and improving code organization.
